Improved git repository directory detection
Added support for out of source builds where the solution file is not in the repository (CMAKE, etc).
Added support for large projects using multiple git repositories.
Added support for submodules.

The git repository directory is now determined from the current file first, then the solution if there is no current file. When determining the git repository, if the file/solution is in a submodule, that will be used rather than the super project.

The plugin has two types of commands "file" (File Log, File Blame, and File Diff) and "repo" (the rest). The "repo" commands now work whether the solution is in a repository or not as long as there is a currently open file that is in a git repository.

If you happen to be working on a project with a checked in solution file, this plugin will work exactly as before with one improvement. If the solution file is not at the root of the repository, the repository root will now be use instead of assuming the solution file is at the repository root which produced very odd behavior (your changes might not show up in the commit dialog).

Updated the version number.